added dynamic frequnecy prescaler selection based on pwm frquency, and added get functions for pwm frequency, duty and channel state

+++ board_firmware/firmware/src/state/runtime/DIORuntimeConfig.h
@@ -37,7 +37,7 @@ extern "C" {
         /**
          * Frequency of PWM output
          */
-        uint16_t PwmFrequency;
+        uint32_t PwmFrequency;
         
         /**
          * Duty Cycle of PWM output
